Coords изменил формат свойств

+++ ConverterAsgard.Common/Parsing/Models/Coords.cs
@@ -5,21 +5,21 @@ namespace ConverterAsgard.Common.Parsing.Models
     public class Coords
     {
         [JsonProperty("latT")]
-        public int LatT { get; set; }
+        public float LatT { get; set; }
 
         [JsonProperty("latN")]
-        public int LatN { get; set; }
+        public float LatN { get; set; }
 
         [JsonProperty("latS")]
-        public int LatS { get; set; }
+        public float LatS { get; set; }
 
         [JsonProperty("lonT")]
-        public int LonT { get; set; }
+        public float LonT { get; set; }
 
         [JsonProperty("lonW")]
-        public int LonW { get; set; }
+        public float LonW { get; set; }
 
         [JsonProperty("lonE")]
-        public int LonE { get; set; }
+        public float LonE { get; set; }
     }
 }
